requirements.txt
google-adk
yfinance
pandas
python-dotenv
fredapi